Move gen_eq into common folder.
