+++ backends/p4tools/common/CMakeLists.txt
@@ -36,6 +36,7 @@ set(
 
   lib/arch_spec.cpp
   lib/format_int.cpp
+  lib/gen_eq.cpp
   lib/model.cpp
   lib/namespace_context.cpp
   lib/symbolic_env.cpp
