+++ backends/p4tools/common/lib/gen_eq.h
@@ -1,9 +1,9 @@
-#ifndef BACKENDS_P4TOOLS_MODULES_TESTGEN_LIB_GEN_EQ_H_
-#define BACKENDS_P4TOOLS_MODULES_TESTGEN_LIB_GEN_EQ_H_
+#ifndef BACKENDS_P4TOOLS_COMMON_LIB_GEN_EQ_H_
+#define BACKENDS_P4TOOLS_COMMON_LIB_GEN_EQ_H_
 
 #include "ir/ir.h"
 
-namespace P4Tools::P4Testgen {
+namespace P4Tools {
 
 /// Generates an equality on a target expression and a keyset expression, recursing into lists.
 /// This supports fuzzy matching on singleton lists: singleton lists are considered the same as
@@ -13,6 +13,7 @@ class GenEq {
  public:
     static const IR::Expression *equate(const IR::Expression *target, const IR::Expression *keyset);
 
+ private:
     static const IR::Expression *equate(const IR::Expression *target,
                                         const IR::DefaultExpression *keyset);
 
@@ -26,11 +27,10 @@ class GenEq {
     static const IR::Expression *equate(const IR::ListExpression *target,
                                         const IR::ListExpression *keyset);
 
- private:
     /// Convenience method for producing a typed Eq node on the given expressions.
     static const IR::Equ *mkEq(const IR::Expression *e1, const IR::Expression *e2);
 };
 
-}  // namespace P4Tools::P4Testgen
+}  // namespace P4Tools
 
-#endif /* BACKENDS_P4TOOLS_MODULES_TESTGEN_LIB_GEN_EQ_H_ */
+#endif /* BACKENDS_P4TOOLS_COMMON_LIB_GEN_EQ_H_ */
