+++ backends/p4tools/modules/testgen/core/small_step/expr_stepper.cpp
@@ -7,6 +7,7 @@
 #include <vector>
 
 #include "backends/p4tools/common/core/solver.h"
+#include "backends/p4tools/common/lib/gen_eq.h"
 #include "backends/p4tools/common/lib/symbolic_env.h"
 #include "backends/p4tools/common/lib/variables.h"
 #include "ir/declaration.h"
@@ -27,7 +28,6 @@
 #include "backends/p4tools/modules/testgen/lib/continuation.h"
 #include "backends/p4tools/modules/testgen/lib/exceptions.h"
 #include "backends/p4tools/modules/testgen/lib/execution_state.h"
-#include "backends/p4tools/modules/testgen/lib/gen_eq.h"
 #include "backends/p4tools/modules/testgen/options.h"
 
 namespace P4Tools::P4Testgen {
